+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## 1.6.0 - Match , 2023
+## 1.6.1 - March 20, 2023
+
+- Fixed Issue with ads being enabled on Web.
+- Added Day Counter to the bottom of the graph
+- Gave background to weight dibits in graph, to not be overlapped with overflowing lines.
+
+## 1.6.0 - March 20, 2023
 
 - Added Graph Screen